+++ packages/illustrations/src/__generated__/spotIcon/data/descriptionMap.ts
@@ -1,0 +1,43 @@
+/**
+ * DO NOT MODIFY
+ * Generated from yarn nx run illustrations:sync
+ */
+
+/**
+ * Mapping of descriptions to associated illustrations.
+ * This is being used on the search portion of the SpotIcon page on the CDS website.
+ * The search query filters the shown illustrations based on matches with name or description.
+ */
+const descriptionMap: Record<string, string[]> = {
+  product: [
+    'productCompliance',
+    'productPro',
+    'productEarn',
+    'productCoinbaseCard',
+    'productWallet',
+  ],
+  icons: ['productCompliance', 'productPro', 'productEarn', 'productCoinbaseCard', 'productWallet'],
+  icon: ['productCompliance', 'productPro', 'productEarn', 'productCoinbaseCard', 'productWallet'],
+  small: ['productCompliance', 'productPro', 'productEarn', 'productCoinbaseCard', 'productWallet'],
+  coinbase: [
+    'productCompliance',
+    'productPro',
+    'productEarn',
+    'productCoinbaseCard',
+    'productWallet',
+  ],
+  '32x32': [
+    'productCompliance',
+    'productPro',
+    'productEarn',
+    'productCoinbaseCard',
+    'productWallet',
+  ],
+  compliance: ['productCompliance'],
+  pro: ['productPro'],
+  earn: ['productEarn'],
+  card: ['productCoinbaseCard'],
+  wallet: ['productWallet'],
+};
+
+export default descriptionMap;