+++ packages/illustrations/src/__generated__/spotIcon/data/names.ts
@@ -1,0 +1,20 @@
+/**
+ * DO NOT MODIFY
+ * Generated from yarn nx run illustrations:sync
+ */
+
+import type { SpotIconName } from '../types/SpotIconName';
+
+/**
+ * An array of all SpotIcon illustrations.
+ * This is being used to display a sheet of all SpotIcon illustration on the CDS website.
+ */
+const names: SpotIconName[] = [
+  'productCoinbaseCard',
+  'productCompliance',
+  'productEarn',
+  'productPro',
+  'productWallet',
+];
+
+export default names;