+++ packages/illustrations/src/__generated__/spotIcon/svgJs/dark/productWallet-0.js
@@ -1,0 +1,3 @@
+module.exports = {
+  content: `<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 32 32"><path fill="#44C28D" d="M29 2H2v18.03h27V2Z"/><path fill="#588AF5" d="M22.736 20.03V29L2 20.03h20.736Z"/><path fill="#0A0B0D" d="M22.736 10.97v9.06H2V2l20.736 8.97Z"/></svg>`,
+};