+++ packages/illustrations/src/__generated__/spotIcon/types/SpotIconName.ts
@@ -1,0 +1,11 @@
+/**
+ * DO NOT MODIFY
+ * Generated from yarn nx run illustrations:sync
+ */
+
+export type SpotIconName =
+  | 'productCoinbaseCard'
+  | 'productCompliance'
+  | 'productEarn'
+  | 'productPro'
+  | 'productWallet';