+++ packages/illustrations/src/__generated__/spotRectangle/types/SpotRectangleName.ts
@@ -142,6 +142,7 @@ export type SpotRectangleName =
   | 'readyToTrade'
   | 'referralsBitcoin'
   | 'referralsBonus'
+  | 'referralsCoinbaseOne'
   | 'retailUSDCRewards'
   | 'scanCode'
   | 'secureAndTrusted'